requirements.txt
numpy==1.26.4 
opencv-python==4.9.0.80 
pillow==10.3.0 
ttkthemes
moviepy
sounddevice
soundfile
